requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
python-multipart
requests
pandas
streamlit
